Added permissions for devices and simulators

+++ controllers/device.js
@@ -4,11 +4,16 @@ const sequelize = new Sequelize(pgURI);
 const Op = Sequelize.Op;
 const DeviceService = require("../services/device");
 const OfficeService = require("../services/office");
+const UserService = require("../services/user");
 const validateGetDevices = require("../validation/device").validateGetDevices;
 const validateDeviceInput = require("../validation/device").validateDeviceInput;
 
 module.exports = {
   getDevices: async function(req, res) {
+    var canGetDevice = await UserService.canGetDeviceAndSimulator(req.user);
+    if (!canGetDevice) {
+      return res.status(403).json({ message: "Forbidden" });
+    }
     const { errors, isValid } = validateGetDevices(req.query);
 
     // Check Validation
@@ -40,6 +45,10 @@ module.exports = {
   },
   createDevice: async function(req, res) {
     (async () => {
+      var canCreateDevice = await UserService.canCreateUpdateDeleteDevice(req.user);
+      if (!canCreateDevice) {
+        return res.status(403).json({ message: "Forbidden" });
+      }
       const { errors, isValid } = validateDeviceInput(req.body, true);
 
       // Check Validation
@@ -72,10 +81,6 @@ module.exports = {
       if (!office_exists) {
         return res.status(400).json({ error: "Office doesn't exist" });
       }
-      //   var canCreateGroup = await UserService.canCreateEditDeleteGroup(req.user, req.body.project_id);
-      //   if (!canCreateGroup) {
-      //     return res.status(403).json({ message: "Forbidden" });
-      //   }
 
       var created_device = await DeviceService.createDevice(deviceFields);
       if (created_device) {
@@ -87,6 +92,10 @@ module.exports = {
     })();
   },
   updateDevice: async function(req, res) {
+    var canUpdateDevice = await UserService.canCreateUpdateDeleteDevice(req.user);
+    if (!canUpdateDevice) {
+      return res.status(403).json({ message: "Forbidden" });
+    }
     if (isNaN(req.params.id)) {
       return res.status(400).json({ error: "Device id is not valid number" });
     } else {
@@ -95,12 +104,7 @@ module.exports = {
       if (!isValid) {
         return res.status(400).json(errors);
       }
-      if (req.params.id) {
-        var deviceExists = await DeviceService.checkIfDeviceExistById(req.params.id);
-        if (!deviceExists) {
-          return res.status(404).json({ error: "Device doesn't exist" });
-        }
-      }
+
       if (req.body.office_id) {
         var office_exists = await OfficeService.checkIfOfficeExists(req.body.office_id);
       }
@@ -128,14 +132,10 @@ module.exports = {
       }
 
       (async () => {
-        var deviceExists = await DeviceService.checkIfDeviceExistById(req.params.id);
+        var deviceExists = await DeviceService.checkIfDeviceExistById(req.params.id, true);
         if (!deviceExists) {
           return res.status(403).json({ error: "Device doesn't exist" });
         }
-        // var canCreateGroup = await UserService.canCreateEditDeleteGroup(req.user, group.project_id);
-        // if (!canCreateGroup) {
-        //   return res.status(403).json({ message: "Forbidden" });
-        // }
 
         var updatedDevice = await DeviceService.updateDevice(req.params.id, deviceFields);
         var device = await DeviceService.returnCreatedOrUpdatedDevice(updatedDevice);
@@ -144,15 +144,14 @@ module.exports = {
     }
   },
   deleteDevice: async function(req, res) {
-    var deviceExists = await DeviceService.checkIfDeviceExistById(req.params.id);
+    var canDeleteDevice = await UserService.canCreateUpdateDeleteDevice(req.user);
+    if (!canDeleteDevice) {
+      return res.status(403).json({ message: "Forbidden" });
+    }
+    var deviceExists = await DeviceService.checkIfDeviceExistById(req.params.id, true);
     if (!deviceExists) {
       return res.status(404).json({ error: "Device doesn't exist" });
     }
-
-    //   var canDeleteProject = await UserService.canDeleteProject(req.user, req.params.id);
-    //   if (!canDeleteProject) {
-    //     return res.status(403).json({ message: "Forbidden" });
-    //   }
     var deleteDevice = await DeviceService.deleteDevice(req.params.id);
     if (deleteDevice) {
       return res.status(200).json({ success: "Device removed successfully" });
@@ -161,16 +160,20 @@ module.exports = {
     }
   },
   getDeviceById: async function(req, res) {
+    var canGetDevice = await UserService.canGetDeviceAndSimulator(req.user);
+    if (!canGetDevice) {
+      return res.status(403).json({ message: "Forbidden" });
+    }
     if (isNaN(req.params.id)) {
       return res.status(400).json({ error: "Device id is not valid number" });
     }
     if (req.params.id) {
-      var deviceExists = await DeviceService.checkIfDeviceExistById(req.params.id);
+      var deviceExists = await DeviceService.checkIfDeviceExistById(req.params.id, true);
       if (!deviceExists) {
         return res.status(404).json({ error: "Device doesn't exist" });
       }
     }
-    var device = await DeviceService.getDeviceById(req.params.id);
+    var device = await DeviceService.getDeviceById(req.params.id, true);
     if (device) {
       return res.status(200).json(device);
     } else {
@@ -179,6 +182,10 @@ module.exports = {
   },
   createSimulator: async function(req, res) {
     (async () => {
+      var canCreateSimulator = await UserService.canCreateUpdateDeleteSimulator(req.user);
+      if (!canCreateSimulator) {
+        return res.status(403).json({ message: "Forbidden" });
+      }
       const { errors, isValid } = validateDeviceInput(req.body, false);
 
       // Check Validation
@@ -203,11 +210,6 @@ module.exports = {
 
       deviceFields.simulator = true;
 
-      //   var canCreateGroup = await UserService.canCreateEditDeleteGroup(req.user, req.body.project_id);
-      //   if (!canCreateGroup) {
-      //     return res.status(403).json({ message: "Forbidden" });
-      //   }
-
       var created_device = await DeviceService.createDevice(deviceFields);
       if (created_device) {
         var device = await DeviceService.returnCreatedOrUpdatedDevice(created_device);
@@ -216,5 +218,84 @@ module.exports = {
         res.status(500).json({ error: "An error occured while creating simulator" });
       }
     })();
+  },
+  updateSimulator: async function(req, res) {
+    var canUpdateSimulator = await UserService.canCreateUpdateDeleteSimulator(req.user);
+    if (!canUpdateSimulator) {
+      return res.status(403).json({ message: "Forbidden" });
+    }
+    if (isNaN(req.params.id)) {
+      return res.status(400).json({ error: "Simulator id is not valid number" });
+    } else {
+      const { errors, isValid } = validateDeviceInput(req.body, false);
+      // Check Validation
+      if (!isValid) {
+        return res.status(400).json(errors);
+      }
+
+      var deviceFields = {};
+      deviceFields.title = req.body.title;
+      if (req.body.resolution) {
+        deviceFields.resolution = req.body.resolution;
+      }
+      if (req.body.dpi) {
+        deviceFields.dpi = req.body.dpi;
+      }
+      if (req.body.udid) {
+        deviceFields.udid = req.body.udid;
+      }
+      if (req.body.screen_size) {
+        deviceFields.screen_size = req.body.screen_size;
+      }
+      deviceFields.retina = req.body.retina;
+
+      (async () => {
+        var deviceExists = await DeviceService.checkIfDeviceExistById(req.params.id, false);
+        if (!deviceExists) {
+          return res.status(403).json({ error: "Simulator doesn't exist" });
+        }
+
+        var updatedDevice = await DeviceService.updateDevice(req.params.id, deviceFields);
+        var device = await DeviceService.returnCreatedOrUpdatedDevice(updatedDevice);
+        res.status(200).json(device);
+      })();
+    }
+  },
+  getSimulatorById: async function(req, res) {
+    var canGetSimulator = await UserService.canGetDeviceAndSimulator(req.user);
+    if (!canGetSimulator) {
+      return res.status(403).json({ message: "Forbidden" });
+    }
+    if (isNaN(req.params.id)) {
+      return res.status(400).json({ error: "Simulator id is not valid number" });
+    }
+    if (req.params.id) {
+      var deviceExists = await DeviceService.checkIfDeviceExistById(req.params.id, false);
+      if (!deviceExists) {
+        return res.status(404).json({ error: "Simulator doesn't exist" });
+      }
+    }
+    var device = await DeviceService.getDeviceById(req.params.id, false);
+    if (device) {
+      return res.status(200).json(device);
+    } else {
+      return res.status(500).json({ error: "Something went wrong" });
+    }
+  },
+  deleteSimulator: async function(req, res) {
+    var canDeleteSimulator = await UserService.canCreateUpdateDeleteSimulator(req.user);
+    if (!canDeleteSimulator) {
+      return res.status(403).json({ message: "Forbidden" });
+    }
+    var deviceExists = await DeviceService.checkIfDeviceExistById(req.params.id, false);
+    if (!deviceExists) {
+      return res.status(404).json({ error: "Simulator doesn't exist" });
+    }
+    var deleteDevice = await DeviceService.deleteDevice(req.params.id);
+    if (deleteDevice) {
+      return res.status(200).json({ success: "Simulator removed successfully" });
+    } else {
+      return res.status(500).json({ message: "Something went wrong" });
+    }
   }
 };
